Fix code

+++ 0x05-pointers_arrays_strings/7-puts_half.c
@@ -6,7 +6,7 @@
  */
 void puts_half(char *str)
 {
-	int i, j;
+	long int i, j;
 
 	/* find length of string */
 	i = 0;
